Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 int main() {
-    float num1,num2,num3,Average;
-    scanf("%f %f %f",&num1, &num2, &num3);
+    int num1,num2,num3,Average;
+    scanf("%d %d %d",&num1, &num2, &num3);
     Average=num1+num2+num3/3;
-    printf("Average: %.2f\n",Average);
+    printf("Average: %.2d\n",Average);
     return 0;
 }
